refactor: destructured contactsPage client side
